# chore: avoid deduplicating docs ids when not needed